+++ TourGuide/src/main/java/com/openclassrooms/tourguide/testUtils/UserPreferences.java
@@ -1,4 +1,4 @@
-package com.openclassrooms.tourguide.testTools;
+package com.openclassrooms.tourguide.testUtils;
 
 import lombok.Data;
