+++ TourGuide/src/test/java/com/openclassrooms/tourguide/TestRewardsService.java
@@ -17,7 +17,7 @@ import gpsUtil.GpsUtil;
 import gpsUtil.location.Attraction;
 import gpsUtil.location.VisitedLocation;
 import rewardCentral.RewardCentral;
-import com.openclassrooms.tourguide.testTools.InternalTestHelper;
+import com.openclassrooms.tourguide.testUtils.InternalTestHelper;
 import com.openclassrooms.tourguide.service.RewardsService;
 import com.openclassrooms.tourguide.domain.User;
 import com.openclassrooms.tourguide.domain.UserReward;
